requirements.txt
# PyTorch with CUDA support
# Note: Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio

# Core ML/AI dependencies
transformers
diffusers
omegaconf
einops
safetensors
timm
tomesd
torchdiffeq
torchsde
tf-keras

# Image/Video processing
decord
scikit-image
opencv-python
albumentations
imageio[ffmpeg]
pillow
moviepy==2.2.1
retina-face  # Imports as 'from retinaface import RetinaFace'
tensorflow==2.15.0  # TensorFlow 2.16+ supports compute capability 12.0 (Blackwell) 

# Audio processing
librosa
SentencePiece

# Utilities and other dependencies
tensorboard
beautifulsoup4
ftfy
func_timeout
onnxruntime

# Web server and API
fastapi
uvicorn
pydantic

# Basic Python dependencies
numpy
requests
tqdm

# Hugging Face Hub (for model uploads)
huggingface_hub

# Real-time avatar system dependencies
websockets
python-multipart
soundfile
webrtcvad
SpeechRecognition
pyttsx3
pyaudio

# Local LLM and RAG dependencies
ollama
chromadb
sentence-transformers

